fix: updated some incorrect import paths from restructuring
