+++ src/components/layout/Layout.tsx
@@ -1,4 +1,4 @@
-import Nav from "../common/nav/Navbar";
+import Nav from "./nav/Navbar";
 import Header from "./header/Header";
 import Footer from "./footer/Footer";
